test: property + contract + AI eval harness + playwright + vitest + CI
Backend: hypothesis property tests for invoice VAT math and tenant
schema shape (caught a real bug: \$ in regex matched trailing newline,
fixed tenancy regex to \A \Z). Contract test asserts OpenAPI shape.
AI eval harness skeleton at app/ai/evals/. Ruff clean on app+tests.
40 backend tests pass.

Frontend: playwright config with desktop+mobile projects, smoke-shell
E2E scenario covering sidebar 224px + bottom nav at 375px + dark mode.
Vitest config + api-client unit tests (200, 403, 402 paths).

CI: .github/workflows/ci.yml with pre-commit + backend (ruff + pytest
with 60% coverage floor) + frontend (typecheck + vitest, skipped until
package-lock.json lands). M1 lint enforced at pre-commit level.

+++ backend/app/tax/rules/fr.py
@@ -42,7 +42,7 @@ class FranceTaxRule:
                 ],
             )
 
-        vat = int(round(subtotal_minor_units * 0.20))
+        vat = round(subtotal_minor_units * 0.20)
         return TaxComputation(
             subtotal_minor_units=subtotal_minor_units,
             lines=[TaxLine(name="TVA 20%", rate=Decimal("0.20"), amount_minor_units=vat)],
